pyproject.toml
[tool.poetry]
name = "tape"
version = "0.1.0"
description = "LLM-to-LM Interpreter for Enhanced Text-Attributed Graph Representation Learning"
authors = ["Devansh Amin <devanshamin97@gmail.com>"]
readme = "README.md"
repository = "https://github.com/devanshamin/TAPE"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "*"
requests = "*"
tqdm = "*"
click = "^8.1.7"
python-dotenv = "^1.0.1"
gdown = "^5.2.0"
numpy = "*"
ogb = "^1.3.6"
jsonargparse = {extras = ["omegaconf"], version = "^4.29.0"}
instructor = {extras = ["litellm"], version = "^1.3.2"} # LLM inference
diskcache = "^5.6.3" # Caching LLM responses
transformers = "^4.41.2" # LM training/inference
sentence-transformers = "^3.0.1" # LM inference
# LM tokenizer
sentencepiece = "^0.2.0" 
protobuf = "^5.27.1" 
# LM training
datasets = "^2.19.2" 
accelerate = "^0.31.0" 
evaluate = "^0.4.2" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"